Parquet: Support _metadata and _common_metadata, multiple row groups, and a generic readTable; also clean up regioned column sources and fix dictionary writing/symbol tables (#954)

+++ DB/src/main/java/io/deephaven/db/v2/locations/parquet/topage/ToPage.java
@@ -91,7 +91,7 @@ public interface ToPage<ATTR extends Attributes.Any, RESULT> {
      * getDictionary returns null.
      * @apiNote null iff {@link #getDictionary()} is null.
      */
-    default ToPage<Attributes.DictionaryKeys, int[]> getDictionaryKeysToPage() {
+    default ToPage<Attributes.DictionaryKeys, long[]> getDictionaryKeysToPage() {
         return null;
     }
 
@@ -131,7 +131,7 @@ public interface ToPage<ATTR extends Attributes.Any, RESULT> {
         }
 
         @Override
-        public ToPage<Attributes.DictionaryKeys, int[]> getDictionaryKeysToPage() {
+        public ToPage<Attributes.DictionaryKeys, long[]> getDictionaryKeysToPage() {
             return toPage.getDictionaryKeysToPage();
         }
